Move redux stuff to src/store and move AccountModal to PaperIndexHeader/...
