+++ frontend/src/modules/App.tsx
@@ -6,8 +6,8 @@ import { toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import "typeface-roboto";
 import LoadingSpinner from "../components/LoadingSpinner";
-import { getUser } from "./auth/selectors";
-import configureStore from "./store";
+import { getUser } from "../store/auth/selectors";
+import configureStore from "../store";
 const AuthenticatedApp = React.lazy(() => import("./AuthenticatedApp"));
 const UnauthenticatedApp = React.lazy(() => import("./UnauthenticatedApp"));
