+++ frontend/src/modules/UnauthenticatedApp/index.tsx
@@ -6,11 +6,11 @@ import { Route, Switch } from "react-router-dom";
 import { toast } from "react-toastify";
 import api from "../../api";
 import { getRefreshToken } from "../../localStorage";
-import { setUser } from "../auth/actions";
+import { setUser } from "../../store/auth/actions";
 import ForgotPasswordPage from "../auth/ForgotPasswordPage";
 import PasswordlessPage from "../auth/PasswordlessPage";
 import ResetPasswordPage from "../auth/ResetPasswordPage";
-import { getUser } from "../auth/selectors";
+import { getUser } from "../../store/auth/selectors";
 import SignInPage from "../auth/SignInPage";
 import SignUpPage from "../auth/SignUpPage";
 import VerifyAccountPage from "../auth/VerifyAccountPage";
