+++ frontend/src/modules/papers/components/headers/PaperIndexHeader/AccountModal.tsx
@@ -1,47 +1,35 @@
-import React, { useState, useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { RouteComponentProps, withRouter } from "react-router";
-import * as Yup from "yup";
-
-import { UserPatchData } from "backend/src/types/users";
-import { setUser, logOut } from "../actions";
-import { getUser } from "../selectors";
-import { getOwnUser, patchOwnUser } from "../../../api/users";
-
 import {
   Button,
   Dialog,
-  DialogContent,
-  DialogContentText,
   DialogActions,
-  DialogTitle
+  DialogContent,
+  DialogContentText
 } from "@material-ui/core";
-import CustomDialogTitle from "../../../components/dialogs/DialogTitleWithCloseButton";
+import { UserPatchData } from "backend/src/types/users";
+import React, { useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { useHistory } from "react-router";
+import * as Yup from "yup";
+import { patchOwnUser } from "../../../../../api/users";
+import CustomDialogTitle from "../../../../../components/dialogs/DialogTitleWithCloseButton";
 import SimpleForm, {
   FormMetadataType
-} from "../../../components/forms/SimpleForm";
+} from "../../../../../components/forms/SimpleForm";
+import { logOut } from "../../../../../store/auth/actions";
+import { getUser } from "../../../../../store/auth/selectors";
 
-interface Props extends RouteComponentProps {
+interface Props {
   render: any;
 }
 
 const AccountModal: React.FC<Props> = props => {
+  const history = useHistory();
   const dispatch = useDispatch();
   const user = useSelector(getUser);
   const [isOpen, setIsOpen] = useState(false);
   const toggleVisibility = () => setIsOpen(!isOpen);
   const [refreshFlag, setRefreshFlag] = useState(false);
   const toggleRefreshFlag = () => setRefreshFlag(!refreshFlag);
-
-  // useEffect(() => {
-  //   getOwnUser().then(user => {
-  //     console.log("account modal set user");
-  //     dispatch(setUser(user));
-  //   });
-  //   if (!user) {
-  //     props.history.push("/login");
-  //   }
-  // }, [refreshFlag]);
 
   if (!user) {
     return <div />;
@@ -97,7 +85,7 @@ const AccountModal: React.FC<Props> = props => {
             color="primary"
             onClick={async () => {
               dispatch(logOut());
-              props.history.push("/login");
+              history.push("/login");
             }}
           >
             Log out
@@ -109,4 +97,4 @@ const AccountModal: React.FC<Props> = props => {
   );
 };
 
-export default withRouter(AccountModal);
+export default AccountModal;
