+++ frontend/src/types/store.ts
@@ -1,4 +1,4 @@
-import { types as auth } from "../modules/auth";
+import { types as auth } from "../store/auth";
 
 /**
  * Describes the overall shape of the application's Redux store state.
